Add option to pass User public ssh key to Gerrit
Currently CI/CD application allows to specify default user for work with
Gerrit, but for normal workflow we need to set ssh public key via Gerrit
UI for having access to clone projects and upload patches.
It's ok for new users, but for fast start we need way to speicify this
key on configuration step, before deployment start.
This aptch adds new field of ssh key for CI/CD application and for
Gerrit application. Specified key will be automatically added to
specified user after deployment Gerrit service.

In script we add check for each option. If this option is not empty we
add it to list of options, otherwise we ignore it.

Change-Id: I8f10f5153ab30361944437fe27e8842420371363